Fix the Coupon resource as percent_off is now a decimal

+++ src/Stripe.net/Entities/StripeCoupon.cs
@@ -39,7 +39,7 @@
         public string Name { get; set; }
 
         [JsonProperty("percent_off")]
-        public int? PercentOff { get; set; }
+        public decimal? PercentOff { get; set; }
 
         [JsonProperty("redeem_by")]
         [JsonConverter(typeof(StripeDateTimeConverter))]
